Rename "theme-chooser" to "theme-picker"
