+++ src/app/shared/navbar/navbar.html
@@ -9,7 +9,7 @@
   <a md-button class="docs-button" routerLink="components">Components</a>
   <a md-button class="docs-button" routerLink="guides">Guides</a>
   <div class="flex-spacer"></div>
-  <theme-chooser></theme-chooser>
+  <theme-picker></theme-picker>
   <a md-button class="docs-button" href="https://github.com/angular/material2" aria-label="GitHub Repository">
     <img class="docs-github-logo"
       src="../../../assets/img/homepage/github-circle-white-transparent.svg"
