--- src/app/shared/theme-chooser/index.ts
@@ -1,1 +1,0 @@
-export * from './theme-chooser';
