+++ src/app/shared/theme-picker/index.ts
@@ -1,0 +1,1 @@
+export * from './theme-picker';
