+++ src/app/shared/theme-picker/theme-picker.spec.ts
@@ -2,15 +2,15 @@ import {MaterialModule} from '@angular/material';
 import {async, TestBed} from '@angular/core/testing';
 
 import {ThemeStorage} from './theme-storage/theme-storage';
-import {ThemeChooser} from './theme-chooser';
+import {ThemePicker} from './theme-picker';
 import {StyleManager} from '../style-manager';
 
 
-describe('ThemeChooser', () => {
+describe('ThemePicker', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       imports: [MaterialModule],
-      declarations: [ThemeChooser],
+      declarations: [ThemePicker],
       providers: [StyleManager, ThemeStorage]
     });
 
@@ -18,7 +18,7 @@ describe('ThemeChooser', () => {
   }));
 
   it('should install theme based on href', () => {
-    const fixture = TestBed.createComponent(ThemeChooser);
+    const fixture = TestBed.createComponent(ThemePicker);
     const component = fixture.componentInstance;
     const href = 'pink-bluegrey.css';
     spyOn(component.styleManager, 'setStyle');
